// src/app/models/ticket.ts
export interface Ticket {
  ticketId?: number;
  busName: string;
  coachNo: string;
  from: string;
  to: string;
  boardingPoint: Array<string>;
  droppingPoint: Array<string>;
  departureDate: string;
  departureTime: string;
  arrivalDate: string;
  arrivalTime: string;
  seatNo: SeatNo;
  price: number;
  availableSeats?: SeatNo[];
}
export type SeatNo =
  | 'A1'
  | 'A2'
  | 'A3'
  | 'A4'
  | 'B1'
  | 'B2'
  | 'B3'
  | 'B4'
  | 'C1'
  | 'C2'
  | 'C3'
  | 'C4'
  | 'D1'
  | 'D2'
  | 'D3'
  | 'D4'
  | 'E1'
  | 'E2'
  | 'E3'
  | 'E4'
  | 'F1'
  | 'F2'
  | 'F3'
  | 'F4'
  | 'G1'
  | 'G2'
  | 'G3'
  | 'G4'
  | 'H1'
  | 'H2'
  | 'H3'
  | 'H4'
  | 'I1'
  | 'I2'
  | 'I3'
  | 'I4'
  | 'J1'
  | 'J2'
  | 'J3'
  | 'J4';